change msg type from Float64 to Float32

+++ jsk_2015_06_hrp_drc/drc_task_common/scripts/vehicle/car_path_visualizer.py
@@ -16,7 +16,7 @@ from math import pi
 class CarPathVisualizer:
     def __init__(self):
         rospy.init_node("CarPathVisualizer", anonymous=True)
-        rospy.Subscriber("/car_steering_wheel", Float64, self.steering_callback)
+        rospy.Subscriber("/car_steering_wheel", Float32, self.steering_callback)
         self.marker_pub = rospy.Publisher("/car_path_marker", MarkerArray, queue_size=10)
         self.r = rospy.Rate(5)
         self.a = 0.0258676
